discuss-parse/src/test/java/io/seven4x/discuss/parse/DiscussParserTest.java
package io.seven4x.discuss.parse;

import org.junit.jupiter.api.*;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

class DiscussParserTest {

    @Test
    void parse() {

        DiscussParser parser = new DiscussParser();


        ParseResult result = parser.parse("小朋友@王总来，@雷松|  你好。");


        assertThat(result.getUsers()).hasSize(2);

    }

    class Calculator {

        public int add(int a, int b) {
            return a + b;
        }

    }
    @Test
    @DisplayName("1 + 1 = 2")
    void addsTwoNumbers() {
        Calculator calculator = new Calculator();
        assertEquals(2, calculator.add(1, 1), "1 + 1 should equal 2");
    }

    @ParameterizedTest(name = "{0} + {1} = {2}")
    @CsvSource({
            "0,    1,   1",
            "1,    2,   3",
            "49,  51, 100",
            "1,  100, 101"
    })
    void add(int first, int second, int expectedResult) {
        Calculator calculator = new Calculator();
        assertEquals(expectedResult, calculator.add(first, second),
                () -> first + " + " + second + " should equal " + expectedResult);
    }
}